Skiftet til Long i ID-er. La til Eier-tabell. La til migreringsskript

+++ src/main/java/no/nav/registre/inntektsmeldingstub/database/repository/ArbeidsforholdRepository.java
@@ -8,6 +8,6 @@ import java.util.Optional;
 import no.nav.registre.inntektsmeldingstub.database.model.Arbeidsforhold;
 
 @Repository
-public interface ArbeidsforholdRepository extends CrudRepository<Arbeidsforhold, Integer> {
+public interface ArbeidsforholdRepository extends CrudRepository<Arbeidsforhold, Long> {
     Optional<Arbeidsforhold> findByArbeidforholdsId(String arbeidforholdsId);
 }
