+++ src/main/java/no/nav/registre/inntektsmeldingstub/database/repository/GraderingIForeldrePengerRepository.java
@@ -6,5 +6,5 @@ import org.springframework.stereotype.Repository;
 import no.nav.registre.inntektsmeldingstub.database.model.GraderingIForeldrepenger;
 
 @Repository
-public interface GraderingIForeldrePengerRepository extends CrudRepository<GraderingIForeldrepenger, Integer> {
+public interface GraderingIForeldrePengerRepository extends CrudRepository<GraderingIForeldrepenger, Long> {
 }
