+++ src/main/java/no/nav/registre/inntektsmeldingstub/database/repository/UtsettelseAvForeldrepengerRepository.java
@@ -6,5 +6,5 @@ import org.springframework.stereotype.Repository;
 import no.nav.registre.inntektsmeldingstub.database.model.UtsettelseAvForeldrepenger;
 
 @Repository
-public interface UtsettelseAvForeldrepengerRepository extends CrudRepository<UtsettelseAvForeldrepenger, Integer> {
+public interface UtsettelseAvForeldrepengerRepository extends CrudRepository<UtsettelseAvForeldrepenger, Long> {
 }
